Add pylint presubmit check.
All checks that were triggered were disabled in pylintrc. I put two lists, one
of warnings that are worthless and one list with the warnings that should be
enabled in a subsequent change. Since fixing these can be messy, this is left as
an exercise to the reader.

BUG=
TEST=

Review URL: http://codereview.chromium.org/8719003