Add support for ProgramDatabaseFile to Ninja
Currently we ignore the ProgramDatabaseFile option in the linker
portion of the GYP file.  This change will cause the /PDB:<name>
switch to be emitted during the link step as part of ldflags.
Since ldflags are appended to the defaults, they override any
default /PDB switches that are current emitted which is consistent
with our current uses.

BUG=126499
TEST=gyptest-link-pdb
Review URL: https://codereview.chromium.org/11359044