Add support for 'resource_include_dirs' in ninja
The msvs generator supports a property called 'resource_include_dirs',
which overrides 'include_dirs' for rc files.

This patch adds similar support for ninja.  If a .rc file is found in
the target's sources, we will define a new 'resource_includes' ninja
variable.  The rc rule now uses 'resource_includes' instead of just
'includes'.

This is useful for cases where an application depends on a large number
of libraries, and so has a long list of 'include_dirs'.  Since rc.exe
does not support response files, it is very easy to exceed the
command-line limit when building the rc file.

The test uses a new rc file 'hello3.rc', which is just a copy of
'subdir/hello2.rc', except that it includes 'include.h' instead of
'subdir/include.h'.

R=scottmg@chromium.org, scottmg
BUG=

Review URL: https://codereview.chromium.org/56623006