Extract XcodeVersion to be a standalone module function
The _XcodeVersion from XcodeSettings class does not need to be a member
function as it does not use any of the field of the class. Extract it
to be a standalone function at the module level, and all the functions
it calls (that also don't use fields from XcodeSettings).

Keep the _XcodeVersion wrapper function that keeps a cache of the
values returned from Xcode to not degrade performance for clients of
the old API that may depends on that cache.

BUG=None
R=mark@chromium.org, thakis@chromium.org

Review URL: https://codereview.chromium.org/148783005